% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_observed_ifr_sir.R
\name{calculate_observed_ifr_sir}
\alias{calculate_observed_ifr_sir}
\title{Calculate Active/Passive IFR/SIR From True IFR/SIR And Detection Rates}
\usage{
calculate_observed_ifr_sir(
  strata,
  passive_asymptomatic_detection,
  passive_symptomatic_detection
)
}
\arguments{
\item{strata}{A data.frame like object with the columns 'population', 'sir',
and 'ifr'. Any additional columns are treated as descriptions for the strata.
If \code{strata} only contains the columns 'population', 'sir', 'ifr' it's assumed
that the sample data set contains no strata.}

\item{passive_asymptomatic_detection}{The probability of detecting a case
presenting asymptomatically through passive surveillance.}

\item{passive_symptomatic_detection}{The probability of detecting a case
presenting symptomatically through passive surveillance.}
}
\value{
The given \code{strata} data.frame with the additional columns 'theta',
'active_sir', 'passive_sir', 'active_ifr', and 'passive_ifr'.
}
\description{
Calculate Active/Passive IFR/SIR From True IFR/SIR And Detection Rates
}
\examples{
strata <- data.frame(
  population = 1000L,
  sir = 0.15,
  ifr = 0.02
)
calculate_observed_ifr_sir(
  strata,
  passive_asymptomatic_detection = 0.2,
  passive_symptomatic_detection = 0.6
)

}
