% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_parameter_estimates.R
\name{calculate_parameter_estimates}
\alias{calculate_parameter_estimates}
\alias{calculate_parameter_estimates.SeverityEstimateFit}
\alias{calculate_parameter_estimates.list}
\alias{calculate_parameter_estimates.default}
\title{Calculate Key Parameter Estimates}
\usage{
calculate_parameter_estimates(x, ...)

\method{calculate_parameter_estimates}{SeverityEstimateFit}(
  x,
  mean_estimate = TRUE,
  median_estimate = TRUE,
  alpha = 0.05,
  include_description = TRUE,
  ...
)

\method{calculate_parameter_estimates}{list}(
  x,
  mean_estimate = TRUE,
  median_estimate = TRUE,
  alpha = 0.05,
  include_description = TRUE,
  ...
)

\method{calculate_parameter_estimates}{default}(x, ...)
}
\arguments{
\item{x}{A object to calculate fatality ratio statistics from, typically a
\linkS4class{SeverityEstimateFit} S4 object.}

\item{...}{Further arguments passed to other methods.}

\item{mean_estimate}{A single logical indicating if the mean estimate for the
parameters should be included in the 'mean_estimate' column of the returned
\code{data.frame}.}

\item{median_estimate}{A single logical indicating if the median estimate for
the parameters should be included in the 'median_estimate column of the
returned \code{data.frame}.}

\item{alpha}{A numeric of significance levels to return the parameters
confidence intervals for. The columns will be in
'\{lower\verb{/}upper\}\verb{_}\{alpha\}'
format (i.e. 'lower\verb{_}05' and 'upper\verb{_}05' for \code{alpha=0.05}).}

\item{include_description}{A single logical indicating if descriptions of the
key parameters should be included in the return value. Setting this to \code{TRUE}
is handy for ad-hoc work, but setting to \code{FALSE} can be better for production
code. The description will be given in the 'parameter_description' column of
the returned \code{data.frame}.}
}
\value{
\code{calculate_parameter_estimates.SeverityEstimateFit} returns a \code{data.frame}
with the column 'parameter'. Other columns are determined by other
parameters, see above for details.

\code{calculate_parameter_estimates.default} signals an error.
}
\description{
Calculate a summary \code{data.frame} describing the key parameters fitted
including:
\itemize{
\item The mildly/asymptomatic passive detection rate,
\item The severe symptoms passive detection rate, and
\item The active detection rate.
}
}
\examples{
draws <- list(
  active_detection = c(0.85, 0.90, 0.92),
  passive_asymptomatic_detection = c(0.15, 0.20, 0.18),
  passive_symptomatic_detection = c(0.55, 0.60, 0.58)
)
calculate_parameter_estimates(draws, alpha = numeric())

}
