% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model.R
\name{check_model}
\alias{check_model}
\title{Check a \code{SeverityEstimateModel} instance}
\usage{
check_model(model, attribute = NULL, override_warning = TRUE)
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel} S4 object instance
representing a model to check.}

\item{attribute}{The attribute being modified by the function calling this
check. If this attribute is set a warning will be issued to the user letting
them know.}

\item{override_warning}{Whether there should be a warning if \code{attribute} is
already set. This function only checks the length of the attribute to
determine if it is set.}
}
\value{
\code{check_model} returns \code{NULL} and will raise an error if there is an issue.
}
\description{
Perform common checks on a \code{SeverityEstimateModel} instances, namely that it
is an instance of that class and has the attributes needed for the particular
function interacting with it.
}
\keyword{internal}
