% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{check_probability}
\alias{check_probability}
\alias{assert_probability}
\title{Checkmate Extensions For Probabilities}
\usage{
check_probability(x)

assert_probability(x, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link[checkmate]{AssertCollection}}.}
}
\description{
This is a small wrapper around \code{\link[checkmate:checkNumber]{checkmate::check_number()}} specifically for
probabilities for convenience.
}
\keyword{internal}
