% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_array_from_indices.R
\name{construct_array_from_indices}
\alias{construct_array_from_indices}
\title{Create An Array From Vectors Of Indices And Values}
\usage{
construct_array_from_indices(..., target = 1L, dim = NULL, dimnames = NULL)
}
\arguments{
\item{...}{Indices given either as an arbitrary number of equal length
numerics or as a single list containing an arbitrary number of equal length
numerics.}

\item{target}{The value to populate the array values with. This parameter
will be coerced to match the length of the indices given.}

\item{dim}{The dim attribute for the array to be created (like
\code{\link[base:array]{base::array()}}) or if \code{NULL} will be inferred from the max index value along
each dimension provided for \code{...}. If not \code{NULL} then the length must match
that of \code{...}.}

\item{dimnames}{Either a named list with corresponding dimname vectors (like
\code{\link[base:array]{base::array()}}), a character vector to just name the dimnames and use a
default sequence for the dimname values, or \code{NULL} for not assigning dimnames
to the output.}
}
\value{
A array with the same number of dimensions as given by \code{...} populated with
the values from \code{target}.
}
\description{
Create an array from vectors of indices, such as (i,j,k) notation, with
optional values and dimnames for the output array.
}
\details{
Overlapping indices will be summed. For example if this function was given
\code{c(1L, 1L)} and \code{c(1L, 1L)} for \code{...} then the resulting array would be 1 x 1
with the entry being \code{2L}.

When given \code{NULL} or a list for \code{dimnames} that argument has the same
behavior as \code{\link[base:array]{base::array()}}. However, this function also provides the ability
to just provide the dimname names as a character and populate the dimnames
values with a sequence (\code{1L:N}). This third option for the \code{dimnames}
argument is for situations where the output array dimensions is not known in
advance without taking the \code{max} of each index vector given.
}
\keyword{internal}
