% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_prior.R
\name{detection_prior}
\alias{detection_prior}
\title{Helper to extract a detection prior from a severity estimate model.}
\usage{
detection_prior(name, prior, default = NULL)
}
\arguments{
\item{name}{The name to display in the warning describing the type of
detection rate being extracted, e.g. 'active' or 'passive symptomatic'.}

\item{prior}{A beta parameterized numeric vector for the prior distribution.}

\item{default}{A default beta parameterization to use, or if \code{NULL} then the
uniformative prior of \code{c("alpha" = 1.0, "beta" = 2.0)} is used.}
}
\value{
A numeric vector containing containing the parameterization for a beta prior.
}
\description{
Helper to extract a detection prior from a severity estimate model.
}
\keyword{internal}
