% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_severity.R
\name{estimate_severity}
\alias{estimate_severity}
\title{Estimate Severity From A Line List}
\usage{
estimate_severity(
  linelist,
  population,
  surveillance,
  outcome,
  time_period = character(),
  strata = character(),
  population_value = utils::tail(names(population), n = 1L),
  surveillance_reference = NULL,
  outcome_reference = NULL,
  time_period_reference = NULL,
  strata_reference = NULL,
  hazard_std = 3,
  degrees_of_freedom = 1L,
  active_prior = c(alpha = 1, beta = 25),
  passive_asymptomatic_prior = c(alpha = 1, beta = 2),
  passive_symptomatic_prior = c(alpha = 25, beta = 1),
  ...
)
}
\arguments{
\item{linelist}{A \code{data.frame} (or \code{data.frame} extending object like a
\code{tibble}) of the line list data.}

\item{population}{A \code{data.frame} (or \code{data.frame} extending object like a
\code{tibble}) of the population data.}

\item{surveillance}{A character of columns describing the surveillance
methods. The values in this column must be coercible to 'Active' and
'Passive'.}

\item{outcome}{A character of columns describing the outcome of the line list
entry. The values in this column must be coercible to 'Asymptomatic',
'Death', and 'Symptomatic'.}

\item{time_period}{A character of columns describing the time period such as
'week' or 'day'}

\item{strata}{A character of columns describing the attributes to stratify
the data on.}

\item{population_value}{A unit length character corresponding to the column
in the \code{population} data.frame (or data.frame like object) for the population
value.}

\item{surveillance_reference}{Either \code{NULL}, a \code{data.frame}, or a vector of
values to use as a reference for the surveillance dimension. If a vector then
it is expected that \code{surveillance} is unit length.}

\item{outcome_reference}{Either \code{NULL}, a \code{data.frame}, or a vector of
values to use as a reference for the outcome dimension. If a vector then
it is expected that \code{outcome} is unit length.}

\item{time_period_reference}{Either \code{NULL}, a \code{data.frame}, or a vector of
values to use as a reference for the time period dimension. If a vector then
it is expected that \code{time_period} is unit length.}

\item{strata_reference}{Either \code{NULL}, a \code{data.frame}, or a vector of
values to use as a reference for the strata dimension. If a vector then
it is expected that \code{strata} is unit length.}

\item{hazard_std}{A single length numeric greater than zero for the standard
deviation to use in the community hazard prior distribution.}

\item{degrees_of_freedom}{A single length integer greater than zero for the
degrees of freedom to use in the model.}

\item{active_prior}{The parameters for the prior beta distribution for the
active detection probability. Can be specified as 'alpha'/'beta',
'mean'/'var', 'mean'/'sd', or 'mean'/'concentration'.}

\item{passive_asymptomatic_prior}{The parameters for the prior beta
distribution for the passive asymptomatic detection probability. Can be
specified as 'alpha'/'beta', 'mean'/'var', 'mean'/'sd', or
'mean'/'concentration'.}

\item{passive_symptomatic_prior}{The parameters for the prior beta
distribution for the passive symptomatic detection probability. Can be
specified as 'alpha'/'beta', 'mean'/'var', 'mean'/'sd', or
'mean'/'concentration'.}

\item{...}{Further optional args that are eventually given to
\code{\link[rstan:sampling]{rstan::sampling()}} related to fitting.}
}
\value{
A \linkS4class{SeverityEstimateFit} S4 object.
}
\description{
Fit a severity estimate model accounting for under reporting of asymptomatic
and mildly symptomatic cases.
}
