% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_extract_data_frame.R
\name{format_extract_data_frame}
\alias{format_extract_data_frame}
\title{Format And Validate A Extract \code{data.frame}s}
\usage{
format_extract_data_frame(
  df,
  name,
  valid_levels,
  level_converter = identity,
  case_insensitive_levels = TRUE
)
}
\arguments{
\item{df}{A \code{data.frame} (or \code{data.frame} extending object like a \code{tibble})
describing the levels.}

\item{name}{A length one character of a name to provide for error messages.}

\item{valid_levels}{A list of characters describing sets of valid levels.
Each character in this list should be the same length.}

\item{level_converter}{A function to process the values in \code{df} before
converting to a factor.}

\item{case_insensitive_levels}{A length one logical indicating if levels are
case insensitive.}
}
\value{
A \code{data.frame} that is similar to the input given, but with a single column
as a factor with:
\itemize{
\item 'Active' and 'Passive' levels for \code{format_surveillance_data_frame}, and
\item 'Asymptomatic', 'Death', and 'Symptomatic' levels for
\code{format_outcome_data_frame}.
}
}
\description{
Format the labels of a single column surveillance/outcome \code{data.frame} for
internal use.
}
\keyword{internal}
