% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidents_population_arrays.R
\name{incidents_population_arrays}
\alias{incidents_population_arrays}
\title{Convert Incidents And Populations To Like Shaped Arrays}
\usage{
incidents_population_arrays(
  linelist,
  population,
  time_period,
  strata,
  surveillance,
  outcome,
  population_value,
  time_period_reference,
  strata_reference,
  surveillance_reference,
  outcome_reference
)
}
\arguments{
\item{linelist}{A line list to convert into an array.}

\item{population}{Population data set to convert into an array.}

\item{time_period}{A character of columns describing the time period such as
'week' or 'day'}

\item{strata}{A character of columns describing the attributes to stratify
the data on.}

\item{surveillance}{A character of columns describing the surveillance
methods.}

\item{outcome}{A character of columns describing the outcome of the line list
entry.}

\item{population_value}{A unit length character corresponding to the column
in the \code{population} data.frame (or data.frame like object) for the population
value.}
}
\value{
A list with names 'incidents', 'population', 'time_period', 'strata',
'surveillance', and 'outcome'. These named list elements correspond to:
\itemize{
\item \code{incidents}: A four dimensional array with the dimensions referring to
\code{time_period}, \code{strata}, \code{surveillance}, and \code{outcome}.
\item \code{population}: A one dimensional array with the dimensions referring to
\code{strata}.
\item \code{time_period}: A data.frame constructed from the \code{time_period} columns
given and whose rownames correspond to the \code{time_period} dimension.
\item \code{strata}: A data.frame constructed from the \code{strata} columns given and
whose rownames correspond to the \code{strata} dimension.
\item \code{surveillance}: A data.frame constructed from the \code{surveillance} columns
given and whose rownames correspond to the \code{surveillance} dimension.
\item \code{outcome}: A data.frame constructed from the \code{outcome} columns given and
whose rownames correspond to the \code{outcome} dimension.
\item \code{linelist_ind}: An integer matrix with the same number of rows as
\code{linelist} and columns corresponding to the dimensions extracted. The values
indicate where each line list element corresponds to in each of those
dimensions.
}

The strata dimensions of \code{incidents} and \code{population} will be the same.
}
\description{
Create an incidents and population arrays that are compatible in shape from
a data.frame like line list and population.
}
\details{
It is expected that all of the \code{time_period}, \code{strata}, \code{surveillance}, and
\code{outcome} columns are all present in \code{linelist}. All of the \code{strata} columns
must be in \code{population}.
}
\keyword{internal}
