% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_levels.R
\name{infer_levels}
\alias{infer_levels}
\title{Infer And Validate Levels For Variables}
\usage{
infer_levels(model, name, name_in, levels = NULL)
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel} S4 object instance
representing a model to check.}

\item{name}{A column name to check for in either the \code{line_list} and/or
\code{population} slots of \code{model}.}

\item{name_in}{A string indication which attribute to check \code{name} against.
Must be 'line_list', 'population', or 'both'.}

\item{levels}{Explicit user provided levels, in order, if provided otherwise
\code{NULL} to infer the levels.}
}
\value{
\code{infer_levels} returns a vector of levels for the particular column given.
Either \code{levels} if non-\code{NULL} or an inferred set of levels.

Explicit \code{levels} may not contain missing or duplicated values. Likewise,
missing values in the source column are rejected before inferred levels are
constructed.
}
\description{
Infer And Validate Levels For Variables
}
\keyword{internal}
