% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data_frame.R
\name{is_data_frame}
\alias{is_data_frame}
\title{Is \code{x} A 'data.frame'?}
\usage{
is_data_frame(x, has_string_columns = character())
}
\arguments{
\item{x}{The object to check and downcast.}

\item{has_string_columns}{A character vector of columns to check for the
presence of as well as check it is a character/factor column.}
}
\value{
Either \code{x} if it is only a 'data.frame' otherwise \code{x} downcasted as a
'data.frame'.
}
\description{
Checks if the given variable is a 'data.frame' like object and if it is not
exactly a 'data.frame' will convert it to one.
}
\keyword{internal}
