% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\name{outcome}
\alias{outcome}
\title{Add An Outcome Severity Mapping To A Model}
\usage{
outcome(
  model,
  name,
  map = c(asymptomatic = "asymptomatic", symptomatic = "symptomatic", severe = "severe")
)
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel}.}

\item{name}{The name of the outcome column, must be present in the \code{linelist}
\code{data.frame} given when initializing the model.}

\item{map}{A named character vector mapping outcome column values to one of
"asymptomatic", "symptomatic", or "severe". The names must be values present
in the outcome column, and the values must be one of "asymptomatic",
"symptomatic", or "severe".}
}
\value{
The \code{model} given modified to contain the outcome severity mapping.
}
\description{
Add An Outcome Severity Mapping To A Model
}
\examples{
line_list <- data.frame(
  patient = 1L:3L,
  week = c(1L, 1L, 2L),
  age = c("Youth", "Adult", "Senior"),
  detection = c("Active", "Passive", "Active"),
  outcome = c("Asymptomatic", "Death", "Symptomatic")
)
population <- data.frame(
  age = c("Youth", "Adult", "Senior"),
  amount = rep(987L, 3L)
)
model <- SeverityEstimateModel(line_list, population) |>
  outcome("outcome", map = c(
    "Asymptomatic" = "asymptomatic",
    "Symptomatic" = "symptomatic",
    "Death" = "severe"
  ))
model

}
