% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passive_from_active_strata.R
\name{passive_from_active_strata}
\alias{passive_from_active_strata}
\title{Calculate The Passively Observed SIR, IFR From Actively Observed SIR, IFR}
\usage{
passive_from_active_strata(
  strata,
  active_detection,
  passive_asymptomatic_detection,
  passive_symptomatic_detection
)
}
\arguments{
\item{strata}{A data.frame like object with the columns 'population', 'sir',
and 'ifr'. Any additional columns are treated as descriptions for the strata.
If \code{strata} only contains the columns 'population', 'sir', 'ifr' it's assumed
that the sample data set contains no strata.}

\item{active_detection}{The probability of detecting a case through active
detection.}

\item{passive_asymptomatic_detection}{The probability of detecting a case
presenting asymptomatically through passive surveillance.}

\item{passive_symptomatic_detection}{The probability of detecting a case
presenting symptomatically through passive surveillance.}
}
\value{
A data.frame of the same structure as \code{strata} but with the 'sir' and 'ifr'
columns adjusted for being passively observed.
}
\description{
Calculate The Passively Observed SIR, IFR From Actively Observed SIR, IFR
}
\keyword{internal}
