% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityEstimateFit-class.R
\name{print.SeverityEstimateFit}
\alias{print.SeverityEstimateFit}
\title{Print Method for \code{SeverityEstimateFit} Objects}
\usage{
\method{print}{SeverityEstimateFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{SeverityEstimateFit}.}

\item{...}{Further arguments passed to the \code{print} method for a \code{stanfit}
object.}
}
\value{
\code{x} invisibly.
}
\description{
Prints a \code{SeverityEstimateFit} object in a structured format. Currently just
prints the \code{model_fit} slot using the \code{print} method for a \code{stanfit} object.
}
