% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityEstimateFit-class.R
\name{print.SummaryEstimateFit}
\alias{print.SummaryEstimateFit}
\title{Print Method for \code{SummaryEstimateFit} Objects}
\usage{
\method{print}{SummaryEstimateFit}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{SummaryEstimateFit}.}

\item{digits}{The number of significant digits to print.}

\item{...}{Further arguments passed to \code{\link[=print.data.frame]{print.data.frame()}}.}
}
\value{
\code{x} invisibly.
}
\description{
Prints a \code{SummaryEstimateFit} object in a structured format.
}
