% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityEstimateModel-class.R
\name{print.SummaryEstimateModel}
\alias{print.SummaryEstimateModel}
\title{Print Method for \code{SummaryEstimateModel} Objects}
\usage{
\method{print}{SummaryEstimateModel}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{SummaryEstimateModel}.}

\item{digits}{The number of significant digits to print for prior
parameters.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Prints a \code{SummaryEstimateModel} object in a structured format.
}
