% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\alias{get_prior}
\title{Internal Prior Setter Helper}
\usage{
prior(model, parameter, ...)

get_prior(model, parameter)
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel}.}

\item{parameter}{The model parameter to get the prior for, should be one
of \code{active}, \code{passive_asymptomatic}, \code{passive_symptomatic}.}

\item{...}{Beta distribution parameterization. Must be one of \code{alpha}/\code{beta},
\code{mean}/\code{var}, \code{mean}/\code{sd}, \code{mean}/\code{concentration}.}
}
\value{
\code{model} modified to contain the prior parameterization for \code{parameter}.

A named numeric vector of \code{alpha} and \code{beta}.
}
\description{
Internal helper used by replacement methods to validate and set prior
parameterizations on a \linkS4class{SeverityEstimateModel}.

Internal helper used by getter methods to extract configured priors from a
\linkS4class{SeverityEstimateModel}, returning a default prior with warning
if unset.
}
\keyword{internal}
