% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\alias{active_prior}
\alias{passive_asymptomatic_prior}
\alias{passive_symptomatic_prior}
\title{Specify Prior Parameterization For A Model Parameter}
\usage{
prior(model, parameter, params)

active_prior(model, params)

passive_asymptomatic_prior(model, params)

passive_symptomatic_prior(model, params)
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel} to specify the prior
distribution for.}

\item{parameter}{The model parameter to specify the prior for, should be one
of 'active', 'passive_asymptomatic', 'passive_symptomatic'.}

\item{params}{A named numeric of length two that represent the
parameterization of a beta distribution. Names must be one of
'alpha'/'beta', 'mean'/'var', 'mean'/'sd'.}
}
\value{
The \code{model} given modified to contain the prior parameterization for
\code{parameter}.
}
\description{
Specify Prior Parameterization For A Model Parameter
}
