% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_reference.R
\name{process_reference}
\alias{process_reference}
\title{Process A Vector Or \code{data.frame} Of Reference Values}
\usage{
process_reference(values, columns)
}
\arguments{
\item{values}{Either \code{NULL}, a \code{data.frame}, or a vector to convert into a
\code{data.frame}.}

\item{columns}{The expected columns of the output reference \code{data.frame}. If
\code{values} is a vector then this can only be a single length vector.}
}
\value{
If \code{values} is \code{NULL} then \code{NULL} otherwise a \code{data.frame} with the columns
specified by \code{columns}.
}
\description{
Convert a given set of values into a formatted \code{data.frame} for reference.
}
\keyword{internal}
