% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_significance_levels.R
\name{process_significance_levels}
\alias{process_significance_levels}
\title{Convert Significance Levels Into Confidence Interval Bounds}
\usage{
process_significance_levels(alpha)
}
\arguments{
\item{alpha}{A numeric of significance levels to consider. Only the first two
decimals are considered and the values in this numeric must be unique.}
}
\value{
A numeric matrix of two rows, with dimnames "lower" and "upper" and
\code{length(alpha)} columns, with dimnames of \code{alpha} formatted as a two digit
number like '05' or '10' for \code{0.05} or \code{0.1}.
}
\description{
Calculate the lower and upper bounds of a confidence interval from a
significance level.
}
\keyword{internal}
