% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{render_template}
\alias{render_template}
\title{Find And Render A Template}
\usage{
render_template(template, template_data = list())
}
\arguments{
\item{template}{The name of the template to render, should match a file under
the \verb{inst/} directory.}

\item{template_data}{The data to use when rendering the template.}
}
\value{
A rendered template as a single length character.
}
\description{
Find And Render A Template
}
\keyword{internal}
