% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityEstimate-package.R
\docType{package}
\name{SeverityEstimate-package}
\alias{SeverityEstimate}
\alias{SeverityEstimate-package}
\title{SeverityEstimate: Bayesian Severity Estimation}
\description{
Bayesian models for severity estimation analysis with the ability to perform analysis by strata as well as incorporating information obtained through active and passive surveillance. Work based on Lessler et al. (2016) \doi{10.1093/aje/kwv452}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://accidda.github.io/SeverityEstimate/}
}

}
\author{
\strong{Maintainer}: Timothy Willard \email{twillard@unc.edu} [copyright holder]

Authors:
\itemize{
  \item Timothy Willard \email{twillard@unc.edu} [copyright holder]
  \item Justin Lessler \email{jlessler@unc.edu} [copyright holder]
}

Other contributors:
\itemize{
  \item ACCIDDA [funder]
}

}
\keyword{internal}
