% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_model.R
\name{stan_model}
\alias{stan_model}
\title{Run A Precompiled Stan Model}
\usage{
stan_model(model_name, ...)
}
\arguments{
\item{model_name}{The name of a model in \code{stanmodels}.}

\item{...}{Further arguments passed to \code{\link[rstan:sampling]{rstan::sampling()}}.}
}
\value{
The output of \code{\link[rstan:sampling]{rstan::sampling()}} called with the given further arguments.
}
\description{
Run A Precompiled Stan Model
}
\keyword{internal}
