% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityEstimateFit-class.R
\name{summary.SeverityEstimateFit}
\alias{summary.SeverityEstimateFit}
\title{Summary Method for \code{SeverityEstimateFit} Objects}
\usage{
\method{summary}{SeverityEstimateFit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{SeverityEstimateFit}.}

\item{...}{Unused.}
}
\value{
\code{summary.SeverityEstimateFit} returns an object of class
\code{SummaryEstimateFit} with elements \code{detection_rates} and
\code{severity_estimates}.
}
\description{
Summarises a fitted severity estimate model by reporting mean detection rate
estimates and mean IFR/SIR estimates by strata.
}
