% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityEstimateModel-class.R
\name{summary.SeverityEstimateModel}
\alias{summary.SeverityEstimateModel}
\title{Summary Method for \code{SeverityEstimateModel} Objects}
\usage{
\method{summary}{SeverityEstimateModel}(object, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{SeverityEstimateModel}.}

\item{...}{Unused.}
}
\value{
\code{summary.SeverityEstimateModel} returns an object of class
\code{SummaryEstimateModel}.
}
\description{
Summarises a user-defined severity estimate model by reporting input data
dimensions, detection probability priors, timestep bounds, mapped detection
and outcome counts, and strata specifications.
}
