% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_summary_estimate_model.R
\name{summary_estimate_model_helpers}
\alias{summary_estimate_model_helpers}
\alias{new_summary_estimate_model}
\alias{format_summary_model_data}
\alias{format_summary_model_priors}
\alias{format_summary_model_timesteps}
\alias{format_summary_model_detection}
\alias{format_summary_model_outcome}
\alias{format_summary_model_map}
\alias{format_summary_model_strata}
\alias{format_summary_model_number}
\alias{print_summary_model_map}
\title{Internal Helpers For \code{SummaryEstimateModel}}
\usage{
new_summary_estimate_model(data, priors, timesteps, detection, outcome, strata)

format_summary_model_data(model)

format_summary_model_priors(model)

format_summary_model_timesteps(model)

format_summary_model_detection(model)

format_summary_model_outcome(model)

format_summary_model_map(model, specification, valid_types)

format_summary_model_strata(model)

format_summary_model_number(x, digits = 3L)

print_summary_model_map(x)
}
\arguments{
\item{data}{A \code{data.frame} summarising model input data dimensions.}

\item{priors}{A \code{data.frame} summarising beta priors.}

\item{timesteps}{A \code{data.frame} summarising the timestep specification.}

\item{detection}{A \code{data.frame} summarising the detection mapping and counts.}

\item{outcome}{A \code{data.frame} summarising the outcome mapping and counts.}

\item{strata}{A \code{data.frame} summarising strata specifications.}

\item{model}{A \linkS4class{SeverityEstimateModel}.}

\item{specification}{A model mapping specification.}

\item{valid_types}{Canonical model types represented in a mapping.}

\item{x}{A numeric vector to format or mapping summary \code{data.frame} to
print.}

\item{digits}{The number of significant digits to keep.}
}
\value{
\code{new_summary_estimate_model} returns an object of class
\code{SummaryEstimateModel}.
}
\description{
Internal helpers used to construct and format \code{SummaryEstimateModel} objects.
}
\keyword{internal}
