% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time}
\alias{time}
\title{Add A Time Specification To A Model}
\usage{
time(model, name, levels = NULL)
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel}.}

\item{name}{The name of the time column, must be present in the \code{linelist}
\code{data.frame} given when initializing the model.}

\item{levels}{The levels for the time or if \code{NULL} inferred from the
\code{linelist} \code{data.frame}. Providing this explicitly is strongly recommended
to ensure the order is as expected by the user.}
}
\value{
The \code{model} given modified to contain the time information.
}
\description{
Add A Time Specification To A Model
}
