% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_map.R
\name{validate_map}
\alias{validate_map}
\title{Validate A Mapping Of Values From A Line List}
\usage{
validate_map(model, name, map, valid_types, required_types = character())
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel} S4 object instance
representing a model to check.}

\item{name}{A column name to check for in either the \code{line_list} slot of
\code{model}.}

\item{map}{A named character vector mapping column values to valid types.}

\item{valid_types}{A character vector of valid types that map values must be.}

\item{required_types}{A character vector of valid types that map values must
include.}
}
\value{
\code{NULL}, if there are no issues.
}
\description{
Validate A Mapping Of Values From A Line List
}
\keyword{internal}
