% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{active_prior}
\alias{active_prior}
\alias{active_prior<-}
\alias{active_prior,SeverityEstimateModel-method}
\alias{active_prior<-,SeverityEstimateModel-method}
\alias{set_active_prior}
\alias{passive_asymptomatic_prior}
\alias{passive_asymptomatic_prior<-}
\alias{passive_asymptomatic_prior,SeverityEstimateModel-method}
\alias{passive_asymptomatic_prior<-,SeverityEstimateModel-method}
\alias{set_passive_asymptomatic_prior}
\alias{passive_symptomatic_prior}
\alias{passive_symptomatic_prior<-}
\alias{passive_symptomatic_prior,SeverityEstimateModel-method}
\alias{passive_symptomatic_prior<-,SeverityEstimateModel-method}
\alias{set_passive_symptomatic_prior}
\title{Get Or Set Model Prior Parameterizations}
\usage{
active_prior(x)

active_prior(x) <- value

\S4method{active_prior}{SeverityEstimateModel}(x)

\S4method{active_prior}{SeverityEstimateModel}(x) <- value

set_active_prior(model, ...)

passive_asymptomatic_prior(x)

passive_asymptomatic_prior(x) <- value

\S4method{passive_asymptomatic_prior}{SeverityEstimateModel}(x)

\S4method{passive_asymptomatic_prior}{SeverityEstimateModel}(x) <- value

set_passive_asymptomatic_prior(model, ...)

passive_symptomatic_prior(x)

passive_symptomatic_prior(x) <- value

\S4method{passive_symptomatic_prior}{SeverityEstimateModel}(x)

\S4method{passive_symptomatic_prior}{SeverityEstimateModel}(x) <- value

set_passive_symptomatic_prior(model, ...)
}
\arguments{
\item{x}{A \linkS4class{SeverityEstimateModel}.}

\item{value}{A named list or named numeric vector providing a beta
parameterization. Must be one of \code{alpha}/\code{beta}, \code{mean}/\code{var}, \code{mean}/\code{sd}.}

\item{model}{A \linkS4class{SeverityEstimateModel}.}

\item{...}{Beta distribution parameterization. Must be one of \code{alpha}/\code{beta},
\code{mean}/\code{var}, \code{mean}/\code{sd}.}
}
\value{
\code{active_prior(x)} returns the current active prior parameterization. If the
active prior has not yet been set, a warning is issued and the default prior
\code{c(alpha = 1.0, beta = 1.0)} is returned.

\code{active_prior(x) <- value} returns \code{x} modified to contain the active prior
parameterization.

\code{set_active_prior(model, ...)} returns \code{model} modified to contain the active
prior parameterization.

\code{passive_asymptomatic_prior(x)} returns the current passive asymptomatic
prior parameterization. If the passive asymptomatic prior has not yet been
set, a warning is issued and the default prior \code{c(alpha = 1.0, beta = 1.0)}
is returned.

\code{passive_asymptomatic_prior(x) <- value} returns \code{x} modified to contain the
passive asymptomatic prior parameterization.

\code{set_passive_asymptomatic_prior(model, ...)} returns \code{model} modified to
contain the passive asymptomatic prior parameterization.

\code{passive_symptomatic_prior(x)} returns the current passive symptomatic prior
parameterization. If the passive symptomatic prior has not yet been set, a
warning is issued and the default prior \code{c(alpha = 1.0, beta = 1.0)} is
returned.

\code{passive_symptomatic_prior(x) <- value} returns \code{x} modified to contain the
passive symptomatic prior parameterization.

\code{set_passive_symptomatic_prior(model, ...)} returns \code{model} modified to
contain the passive symptomatic prior parameterization.
}
\description{
S4 getter and replacement methods for prior parameterizations on a
\linkS4class{SeverityEstimateModel}, plus chainable \verb{set_*} helpers for
pipeline ergonomics.
}
\examples{
line_list <- data.frame(
  patient = 1L:3L,
  week = c(1L, 1L, 2L),
  age = c("Youth", "Adult", "Senior"),
  detection = c("Active", "Passive", "Active"),
  outcome = c("Asymptomatic", "Death", "Symptomatic")
)
population <- data.frame(
  age = c("Youth", "Adult", "Senior"),
  amount = rep(987L, 3L)
)
model <- SeverityEstimateModel(line_list, population) |>
  set_active_prior(mean = 0.9, sd = 0.08) |>
  set_passive_asymptomatic_prior(alpha = 1.0, beta = 1.0) |>
  set_passive_symptomatic_prior(mean = 0.1, var = 0.0064)
model

}
