% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{beta_parameterization}
\alias{beta_parameterization}
\title{Resolve Beta Distribution Parameterization}
\usage{
beta_parameterization(params)
}
\arguments{
\item{params}{A named numeric of length two that represent the
parameterization of a beta distribution. Names must be one of
'alpha'/'beta', 'mean'/'var', 'mean'/'sd'.}
}
\value{
A numeric of length two with names 'alpha' and 'beta' corresponding to the
\code{params} converted to alpha/beta parameterization.
}
\description{
Resolve Beta Distribution Parameterization
}
\keyword{internal}
