% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fatality_ratio.R
\name{calculate_fatality_ratio}
\alias{calculate_fatality_ratio}
\alias{calculate_fatality_ratio.SeverityEstimateFit}
\alias{calculate_fatality_ratio.default}
\title{Calculate Fatality Ratio Statistics}
\usage{
calculate_fatality_ratio(x, ...)

\method{calculate_fatality_ratio}{SeverityEstimateFit}(
  x,
  mean_estimate = TRUE,
  median_estimate = TRUE,
  naive_estimate = FALSE,
  alpha = 0.05,
  ...
)

\method{calculate_fatality_ratio}{default}(x, ...)
}
\arguments{
\item{x}{A object to calculate fatality ratio statistics from, typically a
\linkS4class{SeverityEstimateFit} S4 object.}

\item{...}{Further arguments passed to other methods.}

\item{mean_estimate}{A single logical indicating if the mean estimate for the
ratios should be included in the '\verb{*_}mean\verb{_}estimate' columns of
the returned \code{data.frame}.}

\item{median_estimate}{A single logical indicating if the median estimate for
the ratios should be included in the '\verb{*_}median\verb{_}estimate column
of the returned \code{data.frame}.}

\item{naive_estimate}{A single logical indicating if the naive estimate for
fatality ratio should be included in the 'naive\verb{_}estimate' column.}

\item{alpha}{A numeric of significance levels to return the parameters
confidence intervals for. The columns will be in
'\verb{*_}\{lower\verb{/}upper\}\verb{_}\{alpha\}' format
(i.e. '\verb{*_}lower\verb{_}05' and '\verb{*_}upper\verb{_}05'
for \code{alpha=0.05}).}
}
\value{
\code{calculate_fatality_ratio.SeverityEstimateFit} returns a \code{data.frame}
describing fatality ratios by strata or if now strata were provided to when
fitting a single row \code{data.frame}.

\code{calculate_fatality_ratio.default} signals an error.
}
\description{
Calculate a \code{data.frame} of fatality ratios broken down by strata, as well
as other optional model metrics.
}
