% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_hazard.R
\name{calculate_hazard}
\alias{calculate_hazard}
\alias{calculate_hazard.SeverityEstimateFit}
\alias{calculate_hazard.list}
\alias{calculate_hazard.default}
\title{Calculate Hazard Statistics}
\usage{
calculate_hazard(x, ...)

\method{calculate_hazard}{SeverityEstimateFit}(
  x,
  mean_estimate = TRUE,
  median_estimate = TRUE,
  alpha = 0.05,
  ...
)

\method{calculate_hazard}{list}(
  x,
  time_period,
  strata,
  mean_estimate = TRUE,
  median_estimate = TRUE,
  alpha = 0.05,
  ...
)

\method{calculate_hazard}{default}(x, ...)
}
\arguments{
\item{x}{An object to calculate hazard statistics from, typically a
\linkS4class{SeverityEstimateFit} S4 object.}

\item{...}{Further arguments passed to other methods.}

\item{mean_estimate}{A single logical indicating if the mean estimate for the
hazard should be included in the \code{mean_estimate} column of the returned
\code{data.frame}.}

\item{median_estimate}{A single logical indicating if the median estimate for
the hazard should be included in the \code{median_estimate} column of the
returned \code{data.frame}.}

\item{alpha}{A numeric of significance levels to return the hazard
confidence intervals for. The columns will be in
\verb{\\\{lower/upper\\\}_\\\{alpha\\\}} format
(i.e. \code{lower_05} and \code{upper_05} for \code{alpha=0.05}).}

\item{time_period}{A \code{data.frame} describing the time-period dimension of the
underlying hazard.}

\item{strata}{A \code{data.frame} describing the strata dimension of the
underlying hazard.}
}
\value{
\code{calculate_hazard.SeverityEstimateFit} returns a \code{data.frame} describing
posterior hazard estimates by time period and strata.

\code{calculate_hazard.default} signals an error.
}
\description{
Calculate a \code{data.frame} of posterior infection-hazard estimates broken down
by time period and strata.
}
\keyword{internal}
