% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityEstimateModel-class.R
\name{check_model}
\alias{check_model}
\alias{infer_levels}
\title{Perform Commonly Repeated Checks For Operations On Models}
\usage{
check_model(model, attribute = NULL)

infer_levels(model, name, name_in, levels = NULL, ordered = FALSE)
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel} S4 object instance
representing a model to check.}

\item{attribute}{The attribute being modified by the function calling this
check. If this attribute is set a warning will be issued to the user letting
them know.}

\item{name}{A column name to check for in either the \code{line_list} and/or
\code{population} slots of \code{model}.}

\item{name_in}{A string indication which attribute to check \code{name} against.
Must be 'line_list', 'population', or 'both'.}

\item{ordered}{A boolean indicating if \code{levels} has a specific order. If
\code{TRUE} then \code{levels} cannot be \code{NULL}, users must explicitly indicate what
the ordering is.}
}
\value{
\code{check_model} returns \code{NULL} and will raise an error if there is an issue.

\code{infer_levels} returns a vector of levels for the particular column given.
Either \code{levels} if non-\code{NULL} or an inferred set of levels.
}
\description{
These helpers consolidate common validations and operations on models for use
by functions that operate on models like \code{\link[=strata]{strata()}} or \code{\link[=time]{time()}}.
}
\details{
When using this functionality it is expected that \code{check_model} is always
called first so subsequent helpers will skip validations performed by this
function.
}
\keyword{internal}
