% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate_extensions.R
\name{check_probability}
\alias{check_probability}
\alias{assert_probability}
\alias{assert_bool}
\title{Checkmate Extensions}
\usage{
check_probability(x, na.ok = FALSE, null.ok = FALSE)

assert_probability(
  x,
  na.ok = FALSE,
  null.ok = FALSE,
  .var.name = checkmate::vname(x),
  add = NULL
)

assert_bool(x, na.ok = FALSE, .var.name = checkmate::vname(x))
}
\arguments{
\item{x}{[\code{any}]\cr
Object to check.}

\item{na.ok}{[\code{logical(1)}]\cr
Are missing values allowed? Default is \code{FALSE}.}

\item{null.ok}{[\code{logical(1)}]\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link[checkmate]{AssertCollection}}.}
}
\description{
Convenience wrappers around checkmate for specific internal use cases.
Namely:
\itemize{
\item \code{check_probability}/\code{assert_probability} are wrappers around
\code{\link[checkmate:checkNumber]{checkmate::check_number()}} specifically for probabilities.
\item \code{assert_bool} is a wrappeer around \code{\link[checkmate:checkFALSE]{checkmate::check_false()}} and
\code{\link[checkmate:checkTRUE]{checkmate::check_true()}} for single length logicals.
}
}
\keyword{internal}
