% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_linelist.R
\name{create_sample_linelist}
\alias{create_sample_linelist}
\title{Create A Sample Line List}
\usage{
create_sample_linelist(
  strata,
  times,
  active_detection,
  passive_asymptomatic_detection,
  passive_symptomatic_detection,
  force_of_infection = NULL,
  force_of_infection_mean = -5,
  seed = 1L
)
}
\arguments{
\item{strata}{A data.frame like object with the columns 'population', 'sir',
and 'ifr'. Any additional columns are treated as descriptions for the strata.
If \code{strata} only contains the columns 'population', 'sir', 'ifr' it's assumed
that the sample data set contains no strata.}

\item{times}{A vector denoting time steps for the sample data set.
Must be provided in order.}

\item{active_detection}{The probability of detecting a case through active
detection.}

\item{passive_asymptomatic_detection}{The probability of detecting a case
presenting asymptomatically through passive surveillance.}

\item{passive_symptomatic_detection}{The probability of detecting a case
presenting symptomatically through passive surveillance.}

\item{force_of_infection}{A matrix representing the force of infection where
the row dimension corresponds to the \code{times} given and the column dimension
corresponds to the \code{strata} given or \code{NULL} to generate a random one.}

\item{force_of_infection_mean}{The mean of the initial force of infection to
use. Only used when \code{force_of_infection} is \code{NULL}.}

\item{seed}{The random seed to use for generating the data set.}
}
\value{
A data.frame with the columns 'patient', 'time', 'detection', 'outcome'
along with the strata columns provided in \code{strata}.
}
\description{
Generates a sample line list that can be used as an example data set for
analysis with this package.
}
