% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_linelist.R
\name{create_sample_linelist}
\alias{create_sample_linelist}
\title{Create A Sample Line List}
\usage{
create_sample_linelist(
  strata,
  times,
  active_detection,
  passive_asymptomatic_detection,
  passive_symptomatic_detection,
  case_saturation_from = 500,
  case_saturation_to = 500,
  stochastic_case_saturation = FALSE,
  seed = 1L
)
}
\arguments{
\item{strata}{A data.frame like object with the columns 'population', 'sir',
and 'ifr'. Any additional columns are treated as descriptions for the strata.
If \code{strata} only contains the columns 'population', 'sir', 'ifr' it's assumed
that the sample data set contains no strata.}

\item{times}{A vector denoting time steps for the sample data set.
Must be provided in order.}

\item{active_detection}{The probability of detecting a case through active
detection.}

\item{passive_asymptomatic_detection}{The probability of detecting a case
presenting asymptomatically through passive surveillance.}

\item{passive_symptomatic_detection}{The probability of detecting a case
presenting symptomatically through passive surveillance.}

\item{case_saturation_from}{The lower bound on rates to draw from when
initializing case saturation along the population.}

\item{case_saturation_to}{The upper bound on rates to draw from when
initializing case saturation along the population.}

\item{stochastic_case_saturation}{A single logical indicating if the case
should be stochastic in time or not.
point.}

\item{seed}{The random seed to use for generating the data set.}
}
\value{
A data.frame with the columns 'patient', 'time', 'detection', 'outcome'
along with the strata columns provided in \code{strata}.
}
\description{
Generates a sample line list that can be used as an example data set for
analysis with this package.
}
