% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detection}
\alias{detection}
\title{Add A Detection Type Mapping To A Model}
\usage{
detection(model, name, map = c(active = "active", passive = "passive"))
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel} to specify the prior
distribution for.}

\item{name}{The name of the detection column, must be present in the \code{linelist}
\code{data.frame} given when initializing the model.}

\item{map}{A named character vector mapping detection column values to either
"active" or "passive". The names must be values present in the detection column,
and the values must be either "active" or "passive".}
}
\value{
The \code{model} given modified to contain the detection type mapping.
}
\description{
Add A Detection Type Mapping To A Model
}
\examples{
line_list <- data.frame(
  patient = 1L:3L,
  week = c(1L, 1L, 2L),
  age = c("Youth", "Adult", "Senior"),
  detection = c("Active", "Passive", "Active"),
  outcome = c("Asymptomatic", "Death", "Symptomatic")
)
population <- data.frame(
  age = c("Youth", "Adult", "Senior"),
  amount = rep(987L, 3L)
)
model <- SeverityEstimateModel(line_list, population) |>
  detection("detection", map = c("Active" = "active", "Passive" = "passive"))
model

}
