% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_and_match_data_frame.R
\name{extract_and_match_data_frame}
\alias{extract_and_match_data_frame}
\title{Extract And Match A Subset Of A data.frame}
\usage{
extract_and_match_data_frame(x, cols, subset_x = NULL, stop_on_nomatch = TRUE)
}
\arguments{
\item{x}{A data.frame to extract and match from.}

\item{cols}{A character of columns to extract and match.}

\item{subset_x}{An optional data.frame that is a subset of \code{x} to use for
extracting. If \code{NULL} then is computed from \code{x}.}

\item{stop_on_nomatch}{A single logical indicating if an error should be
thrown in the event of a no match. This can only happen when \code{subset_x} is
given and does not cover the values in \code{x}.}
}
\value{
A list with two named elements, the first is 'df' is the data.frame of the
extracted subset and 'ind' is the match index of \code{x} in 'df'.
}
\description{
Extract a data.frame that is a unique subset of the given data.frame for the
columns and get the index of those rows in the subset.
}
\keyword{internal}
