% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit a severity estimate model instance.}
\usage{
fit(model, ...)
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel} to fit.}

\item{...}{Further optional args that are eventually given to
\code{\link[rstan:sampling]{rstan::sampling()}} related to fitting.}
}
\value{
A \linkS4class{SeverityEstimateFit} S4 object.
}
\description{
Fit a severity estimate model instance.
}
