% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit a severity estimate model instance.}
\usage{
fit(model, ...)
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel} to fit.}

\item{...}{Further optional args that are eventually given to
\code{\link[rstan:sampling]{rstan::sampling()}} related to fitting.}
}
\value{
A \linkS4class{SeverityEstimateFit} S4 object.
}
\description{
Fit a severity estimate model instance.
}
\examples{
\donttest{
set.seed(1)
line_list <- data.frame(
  patient = 1L:3L,
  week = c(1L, 1L, 2L),
  age = c("Youth", "Adult", "Senior"),
  detection = c("Active", "Passive", "Active"),
  outcome = c("Asymptomatic", "Death", "Symptomatic")
)
population <- data.frame(
  age = c("Youth", "Adult", "Senior"),
  amount = rep(987L, 3L)
)
fitted_model <- fit(
  default_model(line_list, population),
  chains = 1L,
  cores = 1L,
  iter = 10L,
  warmup = 5L,
  seed = 1,
  refresh = 0
)
}

}
