% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit a severity estimate model instance.}
\usage{
fit(model, ...)
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel} to specify the prior
distribution for.}

\item{...}{Further optional args that are eventually given to \code{\link[rstan:stan]{rstan::stan()}}
related to fitting.}
}
\description{
Fit a severity estimate model instance.
}
