% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.SeverityEstimateModel}
\alias{fit.SeverityEstimateModel}
\title{Estimate Model Parameters For A Severity Estimate Model}
\usage{
fit.SeverityEstimateModel(object, ...)
}
\arguments{
\item{object}{An instance of \linkS4class{SeverityEstimateModel} whose
parameters to estimate.}

\item{...}{Further optional args that are eventually given to \code{\link[rstan:stan]{rstan::stan()}}
related to fitting.}
}
\description{
Estimate the parameters of a severity estimate model.
}
