% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\alias{inv_logit}
\title{Logit And Inverse Logit}
\usage{
logit(p)

inv_logit(x)
}
\arguments{
\item{p}{A probability to convert to a real value.}

\item{x}{A real value to convert to probability.}
}
\description{
Math utilities for converting between a probability and real valued numbers.
}
\details{
These functions are named to mimic their stan equivalents.
}
\keyword{internal}
