% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityEstimateModel-class.R
\name{print.SeverityEstimateModel}
\alias{print.SeverityEstimateModel}
\title{Print Method for \code{SeverityEstimateModel} Objects}
\usage{
\method{print}{SeverityEstimateModel}(x, ...)
}
\arguments{
\item{x}{An object of class \linkS4class{SeverityEstimateModel}.}

\item{...}{Further arguments passed to the \code{\link[utils:str]{utils::str()}} function.}
}
\value{
\code{x} invisibly.
}
\description{
Prints a \linkS4class{SeverityEstimateModel} object in a structured format.
Currently just a thin wrapper around \code{\link[utils:str]{utils::str()}}.
}
