% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\alias{active_prior}
\alias{passive_asymptomatic_prior}
\alias{passive_symptomatic_prior}
\title{Specify Prior Parameterization For A Model Parameter}
\usage{
prior(model, parameter, ...)

active_prior(model, ...)

passive_asymptomatic_prior(model, ...)

passive_symptomatic_prior(model, ...)
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel} to specify the prior
distribution for.}

\item{parameter}{The model parameter to specify the prior for, should be one
of 'active', 'passive_asymptomatic', 'passive_symptomatic'.}

\item{...}{Beta distribution parameterization. Must be one of 'alpha'/'beta',
'mean'/'var', 'mean'/'sd'.}
}
\value{
The \code{model} given modified to contain the prior parameterization for
\code{parameter}.
}
\description{
Specify Prior Parameterization For A Model Parameter
}
\examples{
line_list <- data.frame(
  patient = 1L:3L,
  week = c(1L, 1L, 2L),
  age = c("Youth", "Adult", "Senior"),
  detection = c("Active", "Passive", "Active"),
  outcome = c("Asymptomatic", "Death", "Symptomatic")
)
population <- data.frame(
  age = c("Youth", "Adult", "Senior"),
  amount = rep(987L, 3L)
)
model <- SeverityEstimateModel(line_list, population) |>
  active_prior(mean = 0.9, sd = 0.08) |>
  passive_asymptomatic_prior(alpha = 1.0, beta = 1.0) |>
  passive_symptomatic_prior(mean = 0.1, var = 0.0064)
model

}
