% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexstanr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{stan_options}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{flexstanr}{\code{\link[flexstanr:stan_options]{stan_options()}}}
}}

