% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityEstimateFit-class.R
\docType{class}
\name{SeverityEstimateFit-class}
\alias{SeverityEstimateFit-class}
\alias{print.SeverityEstimateFit}
\alias{summary.SeverityEstimateFit}
\alias{print.SummaryEstimateFit}
\title{Severity Estimate Fit Class}
\usage{
\method{print}{SeverityEstimateFit}(x, ...)

\method{summary}{SeverityEstimateFit}(object, ...)

\method{print}{SummaryEstimateFit}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \linkS4class{SeverityEstimateFit} or
\code{SummaryEstimateFit}.}

\item{...}{For \code{summary.SeverityEstimateFit()}, unused. For
\code{print.SeverityEstimateFit()}, further arguments passed to the \code{print}
method for a \code{stanfit} object. For \code{print.SummaryEstimateFit()}, further
arguments passed to \code{\link[=print.data.frame]{print.data.frame()}}.}

\item{object}{An object of class \linkS4class{SeverityEstimateFit}.}

\item{digits}{The number of significant digits to print.}
}
\value{
A function-dependent value:
\itemize{
\item \code{summary.SeverityEstimateFit()} returns a \code{SummaryEstimateFit} with
elements \code{detection_rates} and \code{severity_estimates}.
\item \code{print.SeverityEstimateFit()} and \code{print.SummaryEstimateFit()} invisibly
return their input object.
}
}
\description{
This class contains the output from a severity estimate model fitting.
}
\section{Slots}{

\describe{
\item{\code{model_fit}}{A stanfit object returned from fitting a severity estimate
model.}

\item{\code{population}}{The population data used in model fitting in array form with
dimensions corresponding to 'time_period' and 'strata'.}

\item{\code{incidence}}{The line list data used in model fitting in array form
counting incidence with dimensions corresponding to 'time_period', 'strata',
'surveillance', and 'outcome'.}

\item{\code{time_period}}{A data.frame with the variables describing the
'time_period' dimensions of \code{population} and \code{incidence}.}

\item{\code{strata}}{A data.frame with the variables describing the 'strata'
dimensions of \code{population} and \code{incidence}.}

\item{\code{surveillance}}{A data.frame with the variables describing the
'surveillance' dimension of \code{incidence}.}

\item{\code{outcome}}{A data.frame with the variables describing the 'outcome'
dimension of \code{incidence}.}
}}

\section{Functions and methods}{

\itemize{
\item \code{summary(object)} summarises a fitted severity estimate model by reporting
mean detection rate estimates and mean IFR/SIR estimates by strata.
\item \code{print.SeverityEstimateFit(x)} prints a \linkS4class{SeverityEstimateFit}
object in a structured format. Currently this prints the \code{model_fit} slot
using the \code{print} method for a \code{stanfit} object.
\item \code{print.SummaryEstimateFit(x, digits)} prints a \code{SummaryEstimateFit} object
in a structured format.
}
}

\examples{
\donttest{
set.seed(1)
line_list <- data.frame(
  time = c(1L, 1L, 2L),
  age = c("Youth", "Adult", "Senior"),
  detection = c("Active", "Passive", "Active"),
  outcome = c("Asymptomatic", "Death", "Symptomatic")
)
population <- data.frame(
  age = c("Youth", "Adult", "Senior"),
  amount = rep(987L, 3L)
)
model <- default_model(line_list, population)
fitted_model <- suppressWarnings(fit(
  model,
  chains = 1L,
  cores = 1L,
  iter = 10L,
  warmup = 5L,
  seed = 1,
  refresh = 0
))
summary(fitted_model)
}

}
