% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityEstimateFit-class.R
\docType{class}
\name{SeverityEstimateFit-class}
\alias{SeverityEstimateFit-class}
\title{Severity Estimate Fit Class}
\description{
This class contains the output from a severity estimate model fitting.
}
\section{Slots}{

\describe{
\item{\code{model_fit}}{A stanfit object returned from fitting a severity estimate
model.}

\item{\code{population}}{The population data used in model fitting in array form with
dimensions corresponding to 'time_period' and 'strata'.}

\item{\code{incidence}}{The line list data used in model fitting in array form
counting incidence with dimensions corresponding to 'time_period', 'strata',
'surveillance', and 'outcome'.}

\item{\code{time_period}}{A data.frame with the variables describing the
'time_period' dimensions of \code{population} and \code{incidence}.}

\item{\code{strata}}{A data.frame with the variables describing the 'strata'
dimensions of \code{population} and \code{incidence}.}

\item{\code{surveillance}}{A data.frame with the variables describing the
'surveillance' dimension of \code{incidence}.}

\item{\code{outcome}}{A data.frame with the variables describing the 'outcome'
dimension of \code{incidence}.}
}}

