% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeverityEstimateModel-class.R
\docType{class}
\name{SeverityEstimateModel-class}
\alias{SeverityEstimateModel-class}
\alias{SeverityEstimateModel}
\title{Severity Estimate Model}
\usage{
SeverityEstimateModel(line_list, population)
}
\arguments{
\item{line_list}{A line list of cases to model the severity of.}

\item{population}{A dataset containing information on the population broken
down by strataification.}
}
\value{
A \linkS4class{SeverityEstimateModel} S4 object instance representing a model
and its associated metadata.
}
\description{
A representation of a severity estimate model and its metadata.
}
\section{Slots}{

\describe{
\item{\code{line_list}}{A line list of cases to model the severity of.}

\item{\code{population}}{A dataset containing information on the population broken
down by strataification.}

\item{\code{active_prior}}{Parameters for the beta distribution prior for the active
detection rate.}

\item{\code{passive_asymptomatic_prior}}{Parameters for the beta distribution prior
for the passive asymptomatic detection rate.}

\item{\code{passive_symptomatic_prior}}{Parameters for the beta distribution prior
for the passive symptomatic detection rate.}
}}

