% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_model.R
\name{stan_model}
\alias{stan_model}
\title{Run A Stan Model From A Template}
\usage{
stan_model(template, data = list(), fun = rstan::stan, ...)
}
\arguments{
\item{template}{The name of the template to render, should match a file under
the \verb{inst/} directory.}

\item{data}{The data to use when rendering the template.}

\item{fun}{The function to call on the rendered template, typically one of
\code{\link[rstan:stan]{rstan::stan()}} for live use and \code{\link[rstan:stanc]{rstan::stanc()}} for unit testing.}

\item{...}{Further arguments passed to the \code{fun} function.}
}
\value{
The output of \code{fun} called with the given further arguments.
}
\description{
Run A Stan Model From A Template
}
\keyword{internal}
