% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.R
\name{strata}
\alias{strata}
\title{Add A Stratification For IFR/SIR Estimates}
\usage{
strata(model, name, levels = NULL, ordered = FALSE)
}
\arguments{
\item{model}{A \linkS4class{SeverityEstimateModel} to specify the prior
distribution for.}

\item{name}{The name of the strataification column, must be present in both
the \code{linelist} and \code{population} \code{data.frame}s given when initializing the
model.}

\item{levels}{The levels for the stratification or if \code{NULL} inferred from
the \code{linelist}/\code{population} \code{data.frame}s.}

\item{ordered}{Indicator for if the levels are ordered in affect, i.e. age
increasing severity. If \code{TRUE} then \code{levels} must be provided.}
}
\value{
The \code{model} given modified to contain the stratification information.
}
\description{
Add A Stratification For IFR/SIR Estimates
}
